Implement loggers.enableFileTransport config.
