+++ lib/config.js
@@ -70,6 +70,9 @@ config.ensureConfigOverride.fields = [];
  */
 config.loggers = {};
 
+// set to false to disable all file based logging
+config.loggers.enableFileTransport = true;
+
 // transport for console logging
 config.loggers.console = {};
