+++ lib/loggers/index.js
@@ -1,31 +1,19 @@
 /*!
- * Copyright (c) 2012-2019 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2012-2021 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
-const brUtil = require('../util');
-const cc = brUtil.config.main.computer();
 const cluster = require('cluster');
 const config = require('../config');
 const crypto = require('crypto');
+const fileLogger = require('./fileLogger');
 const formatters = require('./formatters');
-const fs = require('fs').promises;
-const path = require('path');
-const uidNumber = require('uid-number');
 const util = require('util');
 const winston = require('winston');
 const WinstonMail = require('winston-mail').Mail;
 const WorkerTransport = require('./WorkerTransport');
 
 const randomBytes = util.promisify(crypto.randomBytes);
-
-// config filenames
-// configured here instead of config.js due to util dependency issues
-cc({
-  'loggers.app.filename': () => path.join(config.paths.log, 'app.log'),
-  'loggers.access.filename': () => path.join(config.paths.log, 'access.log'),
-  'loggers.error.filename': () => path.join(config.paths.log, 'error.log')
-});
 
 // create custom logging levels
 const levels = {
@@ -93,46 +81,12 @@ module.exports = container;
 if(cluster.isMaster) {
   // reserved transports
   container.transports = {
-    console: true,
-    app: true,
-    access: true,
-    error: true,
-    email: true
+    access: null,
+    app: null,
+    console: null,
+    email: null,
+    error: null,
   };
-}
-
-async function chown(filename) {
-  if(config.core.running.userId) {
-    let uid = config.core.running.userId;
-    if(typeof uid !== 'number') {
-      if(process.platform === 'win32') {
-        // on Windows, just get the current UID
-        uid = process.getuid();
-      } else {
-        try {
-          let gid;
-          /* eslint-disable-next-line no-unused-vars */
-          [uid, gid] = await new Promise((resolve, reject) => {
-            uidNumber(uid, (err, uid, gid) => {
-              if(err) {
-                reject(err);
-                return;
-              }
-              resolve([uid, gid]);
-            });
-          });
-        } catch(e) {
-          throw new brUtil.BedrockError(
-            `Unable to convert user "${uid}" to a numeric user id. ` +
-            'Try using a uid number instead.',
-            'Error', {cause: e});
-        }
-      }
-    }
-    if(process.getgid) {
-      await fs.chown(filename, uid, process.getgid());
-    }
-  }
 }
 
 /**
@@ -148,54 +102,21 @@ container.init = async () => {
       ...config.loggers.console,
       format: formatters.fromConfig(config.loggers.console.bedrock),
     });
-    transports.app = new winston.transports.File({
-      ...config.loggers.app,
-      format: formatters.fromConfig(config.loggers.app.bedrock),
-    });
-    transports.access = new winston.transports.File({
-      ...config.loggers.access,
-      format: formatters.fromConfig(config.loggers.access.bedrock),
-    });
-    transports.error = new winston.transports.File({
-      ...config.loggers.error,
-      format: formatters.fromConfig(config.loggers.error.bedrock),
-    });
     transports.email = new WinstonMail(config.loggers.email);
 
-    // set unique names for file transports
-    transports.app.name = 'app';
-    transports.access.name = 'access';
-    transports.error.name = 'error';
-
-    // ensure all files are created and have ownership set to the
-    // application process user
-    const fileLoggers = Object.keys(config.loggers).filter(function(name) {
-      const logger = config.loggers[name];
-      return (brUtil.isObject(logger) && 'filename' in logger);
-    }).map(function(name) {
-      return config.loggers[name];
-    });
-    // TODO: run in parallel
-    for(const fileLogger of fileLoggers) {
-      const dirname = path.dirname(fileLogger.filename);
-      // make directory and chown if requested
-      await fs.mkdir(dirname, {recursive: true});
-      if('bedrock' in fileLogger && fileLogger.bedrock.enableChownDir) {
-        await chown(dirname);
-      }
-      // check file can be opened
-      const f = await fs.open(fileLogger.filename, 'a');
-      await f.close();
-      // always chown log file
-      await chown(fileLogger.filename);
+    if(config.loggers.enableFileTransport) {
+      await fileLogger.init({transports});
     }
 
     // create primary loggers
     for(const cat in config.loggers.categories) {
       const transportNames = config.loggers.categories[cat];
       const options = {levels, transports: []};
-      transportNames.forEach(function(name) {
-        options.transports.push(transports[name]);
+      transportNames.forEach(name => {
+        // only use transports that have been initialized
+        if(transports[name]) {
+          options.transports.push(transports[name]);
+        }
       });
       const logger = winston.createLogger(options);
       const _existingLogger = container.get(cat);
